LayoutTests/http/tests/iframe-monitor/iframe-unload.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script src="./resources/monitor-setup.js"></script>
</head>
<body>
<script>

description("Test iframe with huge resource usage is unloaded.");
window.jsTestIsAsync = true;

onload = async () => {
    if (!await setup()) {
        finishJSTest();
        return;
    }

    // Make sure iframe load is done after rule is set correctly.
    const stage = document.querySelector('#stage');
    const base = 'http://localhost:8080/iframe-monitor/resources';

    stage.innerHTML = `
        <iframe name="frame1" src="http://localhost:8080/iframe-monitor/resources/iframe--eligible--.html"></iframe>
    `;

    await waitUntilUnload('frame1');

    shouldNotBe(`document.querySelector('iframe[name=frame1]').srcdoc`, '""');

    finishJSTest();
}
</script>

<div id="stage"></div>
</body>
</html>
